# Engine/Back/public/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Engine API",
        "description": "API para gerenciar membros da equipe.",
        "contact": {
            "name": "Pablo Kauan M. Timbó",
            "email": "pablokauan.tech@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/publications": {
            "get": {
                "tags": [
                    "Publication"
                ],
                "operationId": "155563eef0fc5315392d5159f6900c7c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Publication"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Publication"
                ],
                "operationId": "1df157695ca57594c4895df075f87360",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Publication"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publication"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/publications/{id}": {
            "get": {
                "tags": [
                    "Publication"
                ],
                "operationId": "79ebde65fcb43bfb28ce7cb8034c2465",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publication"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Publication"
                ],
                "operationId": "25a50d5e76da8405f5d6cacc3be8268c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Publication"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Publication"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Publication"
                ],
                "operationId": "22b8899a0d1ccd1235fd6d44817860b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Publication": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Título do artigo"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Journal"
                    },
                    "autores": {
                        "type": "string",
                        "example": "Silva, A.; Costa, R."
                    },
                    "onde_publicado": {
                        "type": "string",
                        "example": "Revista X"
                    },
                    "ano": {
                        "type": "integer",
                        "example": 2024
                    },
                    "doi": {
                        "type": "string",
                        "example": "10.1234/abcd.2024"
                    }
                },
                "type": "object"
            }
        }
    },
...
}